refactor(rag): change llminference class to base model and move llm to llm_ollama for more clarifivation and update rag_orchestrator
