+++ src/researcher/rag/rag_orchestrator.py
@@ -1,13 +1,13 @@
 from typing import Optional, List
 from pathlib import Path
 from researcher.pipelines import RetrieverPipeline
-from researcher import ( NomicEmbedder, LLMInference
+from researcher import ( NomicEmbedder, LLMOllama
                         , FolderManager, PDFTextExtractor
                     )
 
 class RAGOrchestrator:
     def __init__(self, embedder: NomicEmbedder, retriever: RetrieverPipeline
-                 , llm: LLMInference, fm: FolderManager):
+                 , llm: LLMOllama, fm: FolderManager):
         self.embedder = embedder
         self.retriever = retriever
         self.llm = llm
